BiteBudgetV2/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
Werkzeug==3.0.1
PyJWT==2.8.0
bcrypt==4.1.2
Pillow==10.1.0
pytesseract==0.3.10
opencv-python==4.8.1.78
requests==2.31.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
python-multipart==0.0.6
gunicorn==21.2.0
aiohttp==3.12.14
asyncio-extras==1.3.2
pytest==7.4.3
pytest-flask==1.3.0
